Splits some functions into their own modules.
- Replaces the front-matter module with something simpler.
- Runs the build in the same process as the dev server to
  avoid heavy requires more than once (jsdom).
- Pass the URL as a parameter rather than by the
  environment to the renderer. It's not perfect (marked is
  a singleton) but better than a mixture.
