+++ package.json
@@ -20,7 +20,6 @@
     "cheerio": "1.0.0-rc.3",
     "cpy": "7.3.0",
     "cssnano": "4.1.10",
-    "front-matter": "3.0.2",
     "handlebars": "4.1.2",
     "highlight.js": "9.15.9",
     "marked": "0.7.0",
