+++ scripts/build.js
@@ -1,10 +1,9 @@
 'use strict';
 
 const blogEngine = require('..');
-const baseUrl = process.argv[2];
-const devMode = process.argv.includes('--dev');
+const baseUrl = process.argv[2] || process.env.URL;
 
-blogEngine.build(baseUrl, devMode).catch(e => {
+blogEngine.build(baseUrl).catch(e => {
   console.error(e); // eslint-disable-line no-console
   process.exit(1);
 });
